Given a matrix of integers A of size N x M consisting of 0 and 1. A group of connected 1's forms an island.
From a cell (i, j) such that A[i][j] = 1 you can visit any cell that shares a corner with (i, j) and value in that cell is 1.  More formally, from any cell (i, j) if A[i][j] = 1 you can visit:  (i-1, j) if (i-1, j) is inside the matrix and A[i-1][j] = 1. (i, j-1) if (i, j-1) is inside the matrix and A[i][j-1] = 1. (i+1, j) if (i+1, j) is inside the matrix and A[i+1][j] = 1. (i, j+1) if (i, j+1) is inside the matrix and A[i][j+1] = 1. (i-1, j-1) if (i-1, j-1) is inside the matrix and A[i-1][j-1] = 1. (i+1, j+1) if (i+1, j+1) is inside the matrix and A[i+1][j+1] = 1. (i-1, j+1) if (i-1, j+1) is inside the matrix and A[i-1][j+1] = 1. (i+1, j-1) if (i+1, j-1) is inside the matrix and A[i+1][j-1] = 1. Return the number of islands.